+++ src/components/reception/MultiBarcode.vue
@@ -1,0 +1,282 @@
+<template>
+  <div class="flex flex-col w-full">
+    <traction-heading level="4" :show-border="true">Scan barcodes</traction-heading>
+    <traction-field-group
+      label="Barcodes"
+      attribute="barcodes"
+      for="barcodes"
+      description="The list of labware barcodes to import from source"
+      layout="spacious"
+    >
+      <textarea
+        id="barcodes"
+        v-model="barcodes"
+        placeholder="Scan barcodes to print/import..."
+        rows="4"
+        max-rows="10"
+        name="barcodes"
+        class="w-full text-base py-2 px-3 border border-gray-300 bg-white rounded-md"
+      />
+    </traction-field-group>
+  </div>
+  <div v-if="enablePrinting" id="print" class="p-2 bg-gray-100 rounded p-3">
+    <traction-heading level="4" :show-border="true">Print labels</traction-heading>
+    <div class="flex flex-row space-x-4 w-full">
+      <div class="flex flex-col w-1/2 text-left space-y-2">
+        <traction-label>Barcodes</traction-label>
+        <textarea
+          id="print-barcodes"
+          v-model="printBarcodes"
+          class="w-full py-2 px-3 border border-gray-300 bg-white rounded-md shadow-sm focus:outline-none focus:ring-sdb-100 focus:border-sdb-100 disabled:opacity-75 disabled:bg-gray-200 disabled:cursor-not-allowed"
+          rows="4"
+          disabled
+          max-rows="10"
+        />
+      </div>
+      <DataFetcher :fetcher="fetchPrinters">
+        <div class="flex flex-col text-left w-full space-y-2">
+          <traction-label>Printers</traction-label>
+          <traction-select
+            id="printer-choice"
+            v-model="printerName"
+            data-attribute="printer-options"
+            :options="printerOptions"
+            value-field="text"
+            required
+          />
+        </div>
+      </DataFetcher>
+    </div>
+    <div class="flex justify-end mt-3 w-full">
+      <traction-button
+        id="print-button"
+        class="grow"
+        theme="print"
+        :disabled="!printEnabled"
+        @click="printLabels"
+        >Print Labels</traction-button
+      >
+    </div>
+  </div>
+  <div class="bg-gray-100 rounded p-3">
+    <traction-heading level="4" :show-border="true"> Summary </traction-heading>
+    <div class="flex flex-col w-full">
+      <traction-muted-text class="text-left">Imports data into Traction</traction-muted-text>
+      <p id="importText" class="text-left">
+        Import {{ labwareData.foundBarcodes.size }} labware into {{ props.pipeline }} from
+        {{ reception.text }}
+      </p>
+      <div class="flex flex-row space-x-8 mt-5">
+        <traction-button
+          id="reset"
+          full-width
+          theme="reset"
+          data-action="reset-form"
+          @click="reset"
+        >
+          Reset
+        </traction-button>
+        <traction-button
+          id="importLabware"
+          :disabled="isDisabled"
+          full-width
+          theme="create"
+          data-action="import-labware"
+          @click="importLabware"
+        >
+          <traction-spinner v-if="isFetching" class="h-5"></traction-spinner>
+          <span v-else class="button-text">Import</span>
+        </traction-button>
+      </div>
+    </div>
+  </div>
+</template>
+
+<!-- 
+   A standard form for importing multiple barcodes into traction.
+   Used by lib/receptions/index.js when defining the method of importing barcodes in the reception.  
+-->
+<script setup>
+import { ref, computed, watch } from 'vue'
+import { usePrintingStore } from '@/stores/printing.js'
+import useAlert from '@/composables/useAlert.js'
+import useRootStore from '@/stores'
+import { createBarcodeLabels, createBasicTubeBarcodeLabel } from '@/lib/LabelPrintingHelpers.js'
+import { createReceptionResource, createMessages } from '@/services/traction/Reception.js'
+import { getCurrentDate } from '@/lib/DateHelpers.js'
+import DataFetcher from '@/components/DataFetcher.vue'
+
+const props = defineProps({
+  pipeline: {
+    type: String,
+    default: '',
+  },
+  reception: {
+    type: Object,
+    default: () => ({}),
+  },
+  requestOptions: {
+    type: Object,
+    default: () => ({}),
+  },
+})
+const emit = defineEmits(['importStarted', 'importFinished', 'reset'])
+
+const { showAlert } = useAlert()
+const { createPrintJob, printers } = usePrintingStore()
+const api = useRootStore().api.v1
+const receptionRequest = api.traction.receptions.create
+
+const barcodes = ref('')
+const isFetching = ref(false)
+const printerName = ref('')
+// labwareData is used to store the input barcodes as well as the data fetched from sequencescape which is then imported into traction
+let labwareData = ref({
+  inputBarcodes: [], // inputBarcodes is used to store the barcodes that the user has inputted
+  foundBarcodes: new Set(), // foundBarcodes is used to store the barcodes that have been found in sequencescape
+  attributes: [], // attributes is used to store the attributes that have been found in sequencescape
+})
+let debounceTimer = ref(null) // debounce timer for barcode deletion
+
+// printBarcodes is used to display the barcodes that will be printed
+const printBarcodes = computed(() => Array.from(labwareData.value.foundBarcodes).join('\n'))
+// printerOptions is used to display the printers that are available to print to
+const printerOptions = computed(() =>
+  printers('tube').map(({ name }) => ({
+    text: name,
+  })),
+)
+// printEnabled is used to disable the print button if there are no barcodes to print
+const printEnabled = computed(() => printerName.value && printBarcodes.value)
+const barcodeArray = computed(() => [...new Set(barcodes.value.split(/\s/).filter(Boolean))])
+const isDisabled = computed(() => Boolean(barcodeArray.value.length === 0 || isFetching.value))
+const presentRequestOptions = computed(() =>
+  Object.fromEntries(Object.entries(props.requestOptions).filter(([, v]) => v)),
+)
+//displayPrintOptions is used to decide whether print options should be displayed or not
+const enablePrinting = computed(() =>
+  ['SequencescapeTubes', 'SequencescapeMultiplexedLibraries'].includes(props.reception.value),
+)
+
+watch(barcodes, debounceBarcodeFetch, { immediate: true })
+watch(() => props.requestOptions, debounceBarcodeFetch, { deep: true, immediate: true })
+watch(() => props.reception, resetLabwareData, { immediate: true })
+
+function reset() {
+  emit('reset')
+  barcodes.value = ''
+  resetLabwareData()
+}
+
+/*
+  Resets the labwareData
+*/
+function resetLabwareData() {
+  labwareData.value = {
+    foundBarcodes: new Set(),
+    attributes: [],
+    inputBarcodes: [],
+  }
+}
+
+// Function to fetch the printers
+async function fetchPrinters() {
+  if (usePrintingStore().printers().length === 0) {
+    return await usePrintingStore().fetchPrinters()
+  } else {
+    return { success: true }
+  }
+}
+
+/*
+  create the labels needed for the print job
+*/
+function createLabels(foundBarcodes, date) {
+  const sourceBarcodeList = Array.from(foundBarcodes)
+  const barcodeItems = sourceBarcodeList.map((barcode) => ({ barcode, date }))
+  return createBarcodeLabels({ barcodeItems, createLabelFn: createBasicTubeBarcodeLabel })
+}
+
+/*
+  Creates the print job and shows a success or failure alert
+*/
+async function printLabels() {
+  const { success, message = {} } = await createPrintJob({
+    printerName: printerName,
+    labels: createLabels(labwareData.value.foundBarcodes, getCurrentDate()),
+    copies: 1,
+  })
+
+  showAlert(message, success ? 'success' : 'danger')
+}
+
+//Debounces the barcodes field so that the barcodes are not fetched on every keypress or deletion
+function debounceBarcodeFetch() {
+  if (debounceTimer.value) {
+    clearTimeout(debounceTimer)
+  }
+  if (barcodeArray.value.length === 0) {
+    labwareData.value.foundBarcodes.clear()
+    isFetching.value = false
+    return
+  }
+  isFetching.value = true
+  debounceTimer.value = setTimeout(async () => {
+    await fetchLabware()
+    isFetching.value = false
+  }, 500)
+}
+
+/*
+  Imports the labware into traction.
+  This function is called when the user presses 'import'
+*/
+async function importLabware() {
+  this.$emit('importStarted', {
+    barcodes: labwareData.value.foundBarcodes.size,
+  })
+  //Creates the reception resource and shows a success or failure alert
+  try {
+    const response = await createReceptionResource(
+      receptionRequest,
+      labwareData.value.foundBarcodes,
+      labwareData.value.attributes,
+    )
+    const messages = createMessages({
+      barcodes: Array.from(labwareData.value.foundBarcodes),
+      response,
+      reception: props.reception.value,
+    })
+
+    // we create a different alert for each message
+    messages.forEach(({ type, text }) => {
+      showAlert(text, type)
+    })
+    emit('importFinished')
+  } catch (e) {
+    console.error(e)
+    showAlert(e, 'danger')
+    emit('importFinished')
+  }
+}
+
+/*
+  Fetches the labware from sequencescape and updates the labwareData
+  This function is called when the user presses 'enter' in the barcodes field
+*/
+async function fetchLabware() {
+  try {
+    //Fetch barcodes from sequencescape
+    const { foundBarcodes, attributes } = await props.reception.fetchFunction({
+      requests: api,
+      barcodes: barcodeArray.value,
+      requestOptions: presentRequestOptions.value,
+    })
+    labwareData.value = { foundBarcodes, attributes, inputBarcodes: barcodeArray.value }
+    emit('importFinished')
+  } catch (e) {
+    console.error(e)
+    showAlert(e.toString(), 'danger')
+  }
+}
+</script>
